refactored sealance code into class
