+++ sdk/src/browser.ts
@@ -53,6 +53,7 @@ import {
     RecordProvider,
 } from "./record-provider.js";
 import { RecordScanner } from "./record-scanner.js";
+import { Sealance } from "./sealance/merkletree-support.js";
 
 // @TODO: This function is no longer needed, remove it.
 async function initializeWasm() {
@@ -182,6 +183,7 @@ export {
     RecordProvider,
     RecordScanner,
     RecordSearchParams,
+    Sealance,
     SolutionJSON,
     SolutionsJSON,
     TransactionJSON,
