+++ sdk/src/constants.ts
@@ -113,3 +113,10 @@ export const PUBLIC_TO_PRIVATE_TRANSFER = new Set([
 ]);
 
 export const RECORD_DOMAIN = "RecordScannerV0";
+
+/**
+ * Zero address on Aleo blockchain
+ * Corresponds to field element 0
+ * Used as padding in Merkle trees and as a sentinel value
+ */
+export const ZERO_ADDRESS = "aleo1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqq3ljyzc";
